add bf16 support for fake quantize inference

+++ awq/quantize/qmodule.py
@@ -72,7 +72,7 @@ class ScaledActivation(nn.Module):
         self.scales = nn.Parameter(scales.data)
 
     def forward(self, x):
-        return self.act(x) / self.scales.view(1, 1, -1).to(x.device)
+        return self.act(x) / self.scales.view(1, 1, -1).to(x.device).to(x.dtype)
 
 
 class WQLinear(nn.Module):
